02/10/2025 : Shop System

+++ Assets/Scripts/MainGame/Classes/Player.cs
@@ -4,9 +4,9 @@ public class Player
     public string PlayerName { get; set; }
     public bool IsFirstTimeOpenGame { get; set; }
     public string FormatCurrency { get; set; }
-    public int TrustPoint { get; set; } 
+    public long TrustPoint { get; set; } 
     public long Capital { get; set; }
-    public int Token { get; set; }
+    public long Token { get; set; }
     public List<int> UnlockedCakes { get; set; } 
     public List<PlayerOwnedObject> Ingredients { get; set; }
     public List<PlayerOwnedObject> Cakes { get; set; }
